requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cffi==1.17.1
chardet==5.2.0
cssselect2==0.7.0
Django==5.1.3
fonttools==4.55.3
gunicorn==23.0.0
packaging==24.2
pillow==11.0.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.0
python-dotenv==1.0.1
reportlab==4.2.5
sqlparse==0.5.2
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2024.2
webencodings==0.5.1
zopfli==0.2.3.post1
